requirements.txt
# Core Flask Framework
Flask==2.3.3
Werkzeug==2.3.7

# CORS Support
Flask-CORS==4.0.0

# JWT Authentication
Flask-JWT-Extended==4.5.3
PyJWT==2.8.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.7

# Fix typing_extensions compatibility
typing_extensions>=4.8.0

# Environment Variables
python-dotenv==1.0.0

# Password Hashing
bcrypt==4.0.1

# HTTP Requests
requests==2.31.0

# Core Data Processing & Machine Learning
numpy>=1.24.3,<2.0.0
pandas>=2.0.3,<3.0.0
Pillow>=10.0.1,<11.0.0
opencv-python>=4.8.1.78,<5.0.0

# TensorFlow/Keras for AI Model (Updated for compatibility)
tensorflow>=2.16.1,<3.0.0

# Image Processing
scikit-image>=0.21.0,<1.0.0
matplotlib>=3.7.2,<4.0.0

# Security & Utilities
itsdangerous>=2.1.2,<3.0.0
cryptography>=41.0.7,<42.0.0
click>=8.1.7,<9.0.0

# Date/Time Utilities
python-dateutil>=2.8.2,<3.0.0
pytz>=2023.3,<2024.0.0

# TensorFlow Dependencies (for compatibility)
h5py>=3.8.0,<4.0.0
protobuf>=3.20.0,<5.0.0
six>=1.16.0,<2.0.0

# Production Server
gunicorn>=21.2.0,<22.0.0

# Monitoring & Health Checks
psutil>=5.9.5,<6.0.0

# Optional: API Documentation
flask-restx>=1.2.0,<2.0.0

# Optional: Validation
marshmallow>=3.20.1,<4.0.0
email-validator>=2.0.0,<3.0.0

# Optional: Enhanced Logging
colorlog>=6.7.0,<7.0.0

# Optional: Database Migrations
Flask-Migrate>=4.0.5,<5.0.0
Alembic>=1.12.0,<2.0.0

# Optional: Development Dependencies
pytest>=7.4.2,<8.0.0
pytest-flask>=1.2.0,<2.0.0
pytest-cov>=4.1.0,<5.0.0
black>=23.9.1,<24.0.0
flake8>=6.1.0,<7.0.0

# Optional: Caching
Flask-Caching>=2.1.0,<3.0.0
redis>=5.0.1,<6.0.0

# Optional: Rate Limiting
Flask-Limiter>=3.5.0,<4.0.0

# Optional: Email Support
Flask-Mail>=0.9.1,<1.0.0

# Optional: Async Support
eventlet>=0.33.3,<1.0.0

# Optional: Task Queue
celery>=5.3.4,<6.0.0

# Optional: Monitoring
prometheus-client>=0.17.1,<1.0.0

# Optional: Memory Optimization
memory-profiler>=0.61.0,<1.0.0

# Optional: Regular Expressions
regex>=2023.8.8,<2024.0.0

# PDF Generation
reportlab>=4.0.4,<5.0.0